refactor: move promotion token logic in its own package
Using the latest refactored code, this patch moves the pg_controldata
output parser to the cloudnative-pg/machinery repository and isolate the
promotion token creation logic.
